+++ tests/ui/tests/support/dashboard/config.ts
@@ -1,5 +1,4 @@
 const defaultKymaDashboardAddress = "http://localhost:3001";
-
 const dashboardUrl = (Cypress.env("DOMAIN") as string) || defaultKymaDashboardAddress
 
 export default {
