+++ tests/ui/tests/support/dashboard/index.ts
@@ -1,0 +1,1 @@
+export {default} from "./config"
