+++ tests/ui/tests/support/k8sclient/index.ts
@@ -1,14 +1,6 @@
 import "./namespace"
 import "./service"
-import "./auth"
-import "./loadFile"
 import "./apiRule";
-import {ApiRuleConfig} from "./apiRule";
-import Chainable = Cypress.Chainable;
+export {K8sClient} from "./k8sClient";
+export {KubernetesConfig, getK8sCurrentContext} from "./kubeconfig";
 
-export interface K8sClient {
-    createApiRule(cfg: ApiRuleConfig): Chainable<JQuery>
-    createService(name: string, namespace: string): Chainable<JQuery>
-    createNamespace(name: string): Chainable<JQuery>
-    deleteNamespace(name: string): Chainable<JQuery>
-}
