+++ tests/ui/tests/support/k8sclient/k8sClient.ts
@@ -1,0 +1,9 @@
+import {ApiRuleConfig} from "./apiRule";
+import Chainable = Cypress.Chainable;
+
+export interface K8sClient {
+    createApiRule(cfg: ApiRuleConfig): Chainable<JQuery>
+    createService(name: string, namespace: string): Chainable<JQuery>
+    createNamespace(name: string): Chainable<JQuery>
+    deleteNamespace(name: string): Chainable<JQuery>
+}
