+++ tests/ui/tests/support/k8sclient/service.ts
@@ -1,13 +1,12 @@
 import {loadFixture} from "./loadFile";
 import * as k8s from "@kubernetes/client-node";
-import {post, postApiEndpoint, postApisEndpoint} from "./httpClient";
-import config from "../../config";
+import {postApi} from "./httpClient";
 
 Cypress.Commands.add('createService', (name: string, namespace: string) => {
   cy.wrap(loadFixture('service.yaml')).then((s: k8s.V1Service) => {
     s.metadata.name = name
     // We have to use cy.wrap, since the post command uses a cy.fixture internally
-    cy.wrap(postApiEndpoint(`v1/namespaces/${namespace}/services`, s)).should("be.true");
+    cy.wrap(postApi(`v1/namespaces/${namespace}/services`, s)).should("be.true");
   })
 });
